populate method modifications
Some of the configs that already had populate now lost this method (
Trace Frequency and trace sample), but this will be hopefully soon added
from the abstract class. I am mainly pushing now to update my version
because that is probably one of the reasons why it doesnt work for me
now and i cannot build on top of it.